Refactor Marketplace manager and interface to allow several CatalogSource types installation
Add ConfigMapCatalogSourceReconciler and GRPCImageCatalogSourceReconciler types
that can be used by product reconcilers to create an CatalogSource in
different ways.

All product reconcilers but apicurio-registry have been updated to use
ConfigMapCatalogSourceReconciler as the CatalogSource reconciler to be
used when their corresponding Operator is installed.

The Target.ManifestPackage attribute has been removed from the
marketplace package, as it's not used anymore and it was too specific
of ConfigMap-based CatalogSources